+++ app/core/db_base_model.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Column, Integer, func, DateTime
+from sqlalchemy import Column, BigInteger, func, DateTime
 
 
 class DBLBase:
     """Base model class for all database models."""
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     updated_at = Column(DateTime(timezone=True), onupdate=func.now(), server_default=func.now(), nullable=False)
