+++ app/discussions/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, UniqueConstraint
+from sqlalchemy import Column, BigInteger, String, Text, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
@@ -8,8 +8,8 @@ from app.core.db_base_model import DBLBase
 class Thread(Base, DBLBase):
     __tablename__ = "threads"
 
-    club_id = Column(Integer, ForeignKey("book_clubs.id", ondelete="CASCADE"), nullable=False)
-    author_id = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
+    club_id = Column(BigInteger, ForeignKey("book_clubs.id", ondelete="CASCADE"), nullable=False)
+    author_id = Column(BigInteger, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
     title = Column(String, nullable=False)
     content = Column(Text, nullable=True)
 
@@ -20,8 +20,8 @@ class Thread(Base, DBLBase):
 class Comment(Base, DBLBase):
     __tablename__ = "comments"
 
-    thread_id = Column(Integer, ForeignKey("threads.id", ondelete="CASCADE"), nullable=False)
-    author_id = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
+    thread_id = Column(BigInteger, ForeignKey("threads.id", ondelete="CASCADE"), nullable=False)
+    author_id = Column(BigInteger, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
     content = Column(Text, nullable=False)
 
     author = relationship("User", lazy="selectin")
@@ -33,5 +33,5 @@ class CommentLike(Base, DBLBase):
         UniqueConstraint("comment_id", "user_id", name="uq_comment_likes_comment_id_user_id"),
     )
 
-    comment_id = Column(Integer, ForeignKey("comments.id", ondelete="CASCADE"), nullable=False)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    comment_id = Column(BigInteger, ForeignKey("comments.id", ondelete="CASCADE"), nullable=False)
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
