+++ app/iam/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Column, String, BigInteger, UUID, Integer, DateTime
+from sqlalchemy import Column, String, BigInteger, UUID, DateTime
 
 from app.core.database import Base
 from app.core.db_base_model import DBLBase
@@ -19,6 +19,6 @@ class UserSession(Base, DBLBase):
     __tablename__ = "user_sessions"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     sid_hash = Column(String(64), nullable=False)
     last_used = Column(DateTime(timezone=True), nullable=True)
